Add files via upload
In this analysis, the physical behaviours of cells will be explored before and after being treated with the chemotherapy drug Docetaxel.

This data has been collected over a period of 72 hours which tracks several physical features about each cell being observed. The data was extracted using an advanced microscope which observes the cells label-free and with minimal light.