Add files via upload
Analysing a data set containing the boundary pixel data, that is all the pixels which make up the boundary of the cell. With this new data, new features will be engineered to further describe the cells present. These may be better features in distinguishing between the cells treated and not treated with Docetaxel.